+++ src/pages/ProductDetailPage/index.jsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { useParams } from "react-router-dom";
 
-import { useFetch } from "../../../hooks/useFetch";
+import { useFetch } from "../../hooks/useFetch";
 
-import BoxButton from "../../common/BoxButton";
-import Spinner from "../../common/Spinner";
+import BoxButton from "../../components/common/BoxButton";
+import Spinner from "../../components/common/Spinner";
 import {
   Bottom,
   DetailContainer,
@@ -15,7 +15,7 @@ import {
   Top,
 } from "./styled";
 
-import { BASE_SERVER_URL, PRODUCT_LIST_PATH } from "../../../constants";
+import { BASE_SERVER_URL, PRODUCT_LIST_PATH } from "../../constants";
 
 function ProductDetailPage() {
   const { id: productId } = useParams();
