+++ src/pages/ProductListPage/index.jsx
@@ -1,14 +1,14 @@
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
-import { initializeProductInfoList } from "./../../../stores/productInfoListStore";
-import { useFetch } from "../../../hooks/useFetch";
+import { initializeProductInfoList } from "../../stores/productInfoListStore";
+import { useFetch } from "../../hooks/useFetch";
 
-import Spinner from "../../common/Spinner";
+import Spinner from "../../components/common/Spinner";
 import GridList from "./GridList";
 import ProductCard from "./ProductCard";
 
-import { BASE_SERVER_URL, PRODUCT_LIST_PATH } from "../../../constants";
+import { BASE_SERVER_URL, PRODUCT_LIST_PATH } from "../../constants";
 
 const productListURL = `${BASE_SERVER_URL}${PRODUCT_LIST_PATH}`;
